Migrate to incremental checksum update, get rid of reassembly, add ICMPv6->ICMPv4 translation
This commit is checking in the work done between
e4b8dc405d655d4ac6043a1b7e6151a643b54aeb and
0fe2ffb0d909261bd39c3c766fa4558c59fe5fdd in example-nat46.

Getting rid of L4 identifier on the v6->v4 translation enables
to move to incremental checksum update, which in turn enables
to get rid of reassembly.

Also it adds the skeleton for implementing all of the required
translations, as described by rfc6145.