nat46-core: extend portless protocol support in 1:1 share ratio
Commit 5fed76d2cbafbd9322b92adbfbb5ac116081ecc9 added support for portless protocol
in 1:1 share ratio but it only considered the IPv4 prefix length to check the 1:1
share ratio.
This was too restrictive as it required the assignment of one MAP domain per CE;
therefore take the assigned EA bits length also into account by verifying if all EA
bits are used to generate the IPv4 address in combination with the IPv4 prefix length.
This makes the 1:1 sharing ratio more scalable as no single MAP domain needs to be
assigned per CE.

Signed-off-by: Hans Dedecker <dedeckeh@gmail.com>